backend/requirements-base.txt
beautifulsoup4>=4.12,<5.0
chromadb>=0.5,<1.0
fastapi>=0.115,<1.0
google-auth>=2.35,<3.0
google-cloud-aiplatform>=1.139,<2.0
google-cloud-firestore>=2.20,<3.0
google-genai>=1.0,<2.0
langchain>=0.3,<0.4
langgraph>=0.2,<0.3
Pillow>=10.4,<11.0
pydantic-settings>=2.4,<3.0
pypdf>=5.0,<6.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<0.1
requests>=2.32,<3.0
tavily-python>=0.5,<1.0
transformers>=4.48,<5.0
trafilatura>=1.12,<2.0
uvicorn[standard]>=0.30,<1.0
